manyToMany relationship dazugepackt und cascade types angefangen

+++ three-d-portfolio-server/src/main/java/org/threeDPortfolioGallery/workloads/Theme.java
@@ -13,7 +13,7 @@ public class Theme extends PanacheEntity {
 
     public String mat_inside;
 
-    public  int light_intensity;
+    public  float light_intensity;
 
     public String model_path;
